Photon: Add jetpack_photon_url filter only when photon is active
There are two calls to add_filter('jetpack_photon_url', 'jetpack_photon_url'): one
is in load_jetpack.php, and one is in functions.photon.php. Clean up this
duplication by:

* Removing the add_filter() call in load_jetpack.php.
* Calling add_filter() in functions.photon.php only when the photon module is active.

I chose to put the add_filter() call in functions.photon.php because it looks like
this will be easier to sync with wpcom.

Also, classes from the packages must not be loaded until all of the plugins have
been loaded. Jetpack::is_module_active() uses the Jetpack_Options class. So, put the
add_filter() call in a new function that is hooked to the plugins_loaded action.